fish-speech/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0

# Azure Speech
azure-cognitiveservices-speech==1.34.0

# Data processing
numpy==1.26.4

# Database
sqlalchemy==2.0.27
psycopg2-binary==2.9.9

# Utilities
pydantic==1.10.13
python-multipart==0.0.7
aiofiles==24.1.0
